+++ lib/shadcn_ui.ex
@@ -28,6 +28,7 @@ defmodule ShadcnUI do
     ShadcnUI.Components.Forms.Label,
     ShadcnUI.Components.Forms.NativeSelect,
     ShadcnUI.Components.Forms.RadioGroup,
+    ShadcnUI.Components.Forms.Slider,
     ShadcnUI.Components.Forms.Switch,
     ShadcnUI.Components.Forms.Textarea
   ]